Implement config command v3
