+++ 1.21.5/src/testmod/java/testmod/config/ConfigTestmod.java
@@ -1,11 +1,11 @@
 package testmod.config;
 
-
-import net.minecraft.util.Identifier;
+import net.fabricmc.fabric.api.command.v2.CommandRegistrationCallback;
 import top.offsetmonkey538.monkeylib538.api.ConfigCommandApi;
 import top.offsetmonkey538.offsetconfig538.api.config.ConfigHolder;
 import top.offsetmonkey538.offsetconfig538.api.config.ConfigManager;
 
+import static net.minecraft.server.command.CommandManager.literal;
 import static testmod.Testmod.*;
 
 public class ConfigTestmod {
@@ -14,8 +14,12 @@ public class ConfigTestmod {
         final ConfigHolder<ModConfig> config = ConfigManager.INSTANCE.init(ConfigHolder.create(ModConfig::new, LOGGER::error));
         final ConfigHolder<ModConfig2> config2 = ConfigManager.INSTANCE.init(ConfigHolder.create(ModConfig2::new, LOGGER::error));
 
-        ConfigCommandApi.INSTANCE.register(MOD_ID, config);
-        ConfigCommandApi.INSTANCE.register(Identifier.of(MOD_ID, "2"), config2);
+        CommandRegistrationCallback.EVENT.register((dispatcher, commandRegistryAccess, registrationEnvironment) -> {
+            dispatcher.register(literal(MOD_ID)
+                            .then(ConfigCommandApi.INSTANCE.createConfigCommand("main", config))
+                            .then(ConfigCommandApi.INSTANCE.createConfigCommand("second", config2))
+            );
+        });
 
         LOGGER.info("hello: {}", config.get().hello);
         //LOGGER.info("Number: " + config.Number);
